refactor: move effects from mixins to wallet
